Add: Structured inbox folder and created ChatUserCard component
