+++ src/pages/user/inbox/interface/index.ts
@@ -1,0 +1,8 @@
+export interface IChatUser {
+  id: string;
+  name: string;
+  lastMessage: string;
+  unread: number;
+  time: string;
+  avatar: string | null;
+}
